Fix for hit sorting issue
On rare occasions there can be issues with sorting RPC and MDT hits in the same station.
This fix uses id-based sorting to ensure that RPC and MDT hits are handled correctly, thus avoiding the addition of duplicate hits to the track in the hit recovery, as reported in ATLASRECTS-6410.