Run the TnP analysis also for the truth selection
The TnP selection was only run for the truth selection
This adds a TnP triuth selection driver file and runs the Tnp selection
also now for the truth selection.

In addition, it modified the selection cuts used by the TnP selection to
be the same as that for the standard analyses. In principle we might need
to modify that selection to closer to the analysis that is used for the
full offline analyses.