Remove doTrigMuonConfig flag
Replacing the trigger slice flag doTrigMuonConfig with the corresponding flag that is part of the ConfigFlags (Muon.MuonTrigger). Replaced all teh places that check the flag in the muon configuration, replaced the setting of the flag in runHLT_standalone (it was also being set in MuonRecoSequences.py, but this isn't needed, so was dropped), and removed the flag from the MuonSLiceFlags (ATR-20389).