Fix trigger filter for data overlay
Rewrite the `BSFilter` algorithm to use the `TrigDecisionTool` for
trigger-based event filtering on RAW. This ensure compatibility with
past and future data. Change the `--triggerBit` transform argument from
the CTP ID (240) to the CTP item name ("L1_ZB"), which is guaranteed to
work also if the item ID changes.

Also apply some general code cleanups.

Closes ATLASSIM-5419.