TrigValTools: speedup chainDump script
The `chainDump.py` script (used in all trigger ART tests) was taking
about 1 minute to dump the YAML information of all trigger chains.
Re-write the code to avoid the nested loops over chains and all chain
steps. Now the script takes about 8 seconds for the LS2_v1 menu.

This also fixes a bug in the script extracting the step counts for
chains that start with the same string as another chain.