Athena switched to the external version of GeoModelXml
1. Dropped DetectorDescription/GeoModelXml from the Athena repository
2. Athena clients switched to the external version of GeoModelXml
3. In order to fix a problem related to this migration, which shows up
at Geo2G4 translation, changed the way Geo2G4 determines whether it should
translate a GeoModel volume into a G4 assembly. Previously the check was
performed on both the material pointer and name, and now only the name
comparison is done. This mechanism should probably be revisited for the
long run.