Switched LArHitsTestTool to reading CaloDetDescrManager from the Detector Store
This patch is required for fixing failed ART test, which broke because of the
latest update in LArHitsTestTool. In the long term, though, we need to revisit
the CaloDetDescrManager usage patterns in Simulation in order to determine whether
it is really needed to construct it by the converter at initialization.