TrigConfigSvc: delete TrigConfigSvc service
After the removal of the legacy menu support and the migration of all
clients (e.g. using `xAODConfigSvc`), the `TrigConfigSvc` only
functioned as a wrapper to create the LVL1/HLTConfigSvc. This is no
longer needed as these services are created explicitly where needed.

Also delete obsolete `TrigConfigSvcConfig.py`.

Relates to ATR-23102.