remove AsgTools::setName()
We discussed that a long time ago and I thought I had already done it,
but apparently not: Essentially we do not advertise any workflows in
which the name of components can/ought to be changed after the fact,
and it is liable to cause issues unless strict precautions are taken.
So the solution was/is to just remove AsgTools::setName().