TrigTauRec: fix RNN ID for tracktwoLLP chains
Hello,

As discussed in ATR-24364, the tau trigger chains with LLP-specific RNN tau identification are not configured properly, because the configuration uses the same tool name for Long Lived Particles and regular tau ID.
So we configure the tool according to the first tau chain in the menu that uses RNN ID (it seems it's regular ID being called first), then we reuse this tool configuration. This is fixed by using unique names.

Adding the 'urgent' flag as we'd like to have this in for the upcoming MC sample production expected imminently.

Cheers,
Bertrand