tauRecTools: remove deprecated functionality in TauTrackFinder
Hello,

This MR is removing a deprecated feature in the tau-track association.
In the old days (non-MT R22 trigger), the track particle container was retrieved elsewhere (TrigTauRec) and passed as an argument to the TauTrackFinder execute method to build the tau tracks.
Now, the track particle container is retrieved from within TauTrackFinder, and we no longer need to pass it as an argument.

The reconstruction output is unchanged.

Cheers,
Bertrand