HLTConfigSvc: drop support of legacy menus
Drop the support of reading the Run-1&2 legacy menus in `HLTConfigSvc`
as we now exclusively use these menus via the JSON conversion. With
that the service can no longer be used to access any trigger
information, i.e. it no longer implements `IHLTConfigSvc`. Its sole
purpose is to write the HLT menu into the DetectorStore during
`initialize`.

Relates to ATR-23102.