TauAnalysisTools: drop deprecated energy calibration from tau smearing tool
Hello,

This MR is dropping a number of outdated and deprecated functionalities of the tau smearing tool inherited from R21.

In R22, the MVA tau energy calibration is applied by default in the tau reconstruction, so we should not try to apply MVA TES in TauAnalysisTools (while in R21, MVA TES was not applied by default, so we had to apply it at analysis level).
Similarly for the intermediate "combined TES" calibration, there is no need to run this at analysis level.

There is another functionality to enforce calo-only pt when the calo-only pt differs from MVA pt by a large amount.
Comments have been added to the code to explain the motivation, but it has not been validated in R22, so it should not be applied by default.
Anyway, we should consider a different approach for vetoing muons faking taus in R22, forcing tau pt is very ad-hoc... We should rather move this to a selection tool, or develop a dedicated tool (RNN?) to reject muons.

A small change was made in the combined TES tool: when we can't evaluate the compatibility between calo pt and MVA pt (because the parametrisation is not available), we should not force calo pt.

For completeness, I also updated the final R22 Run2 combined TES calibration file for the MET significance calculation, but this is still not the appropriate thing to do: the MET significance currently uses calo pt resolution.
We need to derive the MVA pt resolution and provide a new calibration file so that it can be used for the MET significance.

Cheers,
Bertrand