Disabling tau thinning in DAOD_PHYS
It is expedient to disable the tau thinning in DAOD_PHYS for two reasons:
- it will allow the tau CP group to use DAOD_PHYS and thus avoid making at least some of their dedicated DAODs
- it will solve the current issue where the AnalysisTau (and hence AnalysisMET) content in PHYSLITE is different depending on whether PHYSLITE is made from AOD (unthinned taus) or DAOD_PHYS (thinned taus)

The cost is 1.5KB/event which is modest given that we will not need to make all of the tau CP DAODs.

This MR comments out the relevant lines to disable tau thinning, and a corresponding line in the ID track thinning to ensure the extra taus keep their associated ID tracks.