TauAnalysisTools: drop deprecated properties for TauSmearingTool
Hello,

This MR is fixing an oversight from !47706 : two properties have disappeared from CommonSmearingTool and this needs to be propagated upstream to TauSmearingTool.
The `ApplyMVATESQualityCheck` default value is also harmonised and changed to false.

Adding the urgent flag as !47706 might cause trouble if people try to run TauSmearingTool.

Cheers,
Bertrand