DerivationFrameworkTau and DerivationFrameworkPhys: update tau content
Hello,

This MR is updating the tau smart slimming list to reflect the final R22 AOD content.

Removed:
* AntiKt4LCTopoJets no longer in AODs (therefore jetLink is useless)
* ptFinalCalib and etaFinalCalib are identical to pt and phi, but used by TauAnalysisTools, so to be on the safe side we keep them for now, but drop phiFinalCalib and mFinalCalib
* ptCombined.etaCombined.phiCombined.mCombined is not needed for analysis
* seedJetWidth.seedJetJvt.seedTrackWidthPt1000.seedTrackWidthPt500 (LCTopo seed jet properties) are not decorated onto the tau in R22, we'll have to revise the strategy for fake tau estimate

Added:
* tau neutral PFOs (neutralPFOLinks.pi0PFOLinks, TauNeutralParticleFlowObjectsAux.pt.eta.phi.m): needed by analyses exploiting tau polarisation
* NNDecayMode is the NN version of PanTau_DecayMode
* ptTauEnergyScale.etaTauEnergyScale possibly used to compute compatibility between calo pt and MVA pt in TauCombinedTES::getUseCaloPtFlag

Minor update to DAOD_PHYS: variables now in the smart slimming list were removed from ExtraVariables. Charged PFOs are no longer in AODs, one has to use TauTracks which are identical physics objects.

Update to PHYSLITE: updated list of variables for AnalysisTauJets, and introduction of TauTracks.

A bug was found in the TauTrackParticleThinning code (when no SelectionString is passed, no TauTrack passes the thinning). This will be fixed in a follow-up MR.

Adding the 'urgent' flag as this MR is needed for the imminent production of DAOD_PHYS for the R22 Run2 reprocessing.

Cheers,
Bertrand