DerivationFrameworkInDet: fix TauTrackParticleThinning when selection string is empty
Hello,

This MR fixes a problem with the thinning of tau tracks when no selectrion string is passed.
In that case, the ID tracks are thinned properly, but the block that thins the tau tracks was not getting any taus to work with (empty `tauToCheck` vector).

Adding the 'urgent' flag as this MR is needed for the imminent DAOD_PHYS production for the R22 Run2 reprocessing.

Cheers,
Bertrand