Add unit test for writing indexed TTree container with diff no objects
The test writes 2 containers in ROOTTREEINDEX technology with different
number of object in each - 3 objects in ContainerA and 1 object in
ContainerB per transaction. There are 2 transactions.
References to the objects are kept and used to verify that the correct
object is retrieved when readingl.
Objects are DbStrings and the strings are written out so the log
comparison can compare if they are OK.
Discussed in ATEAM-788