TriggerJobOpts: fix BFieldFromDCS modifier
Fix the `BFieldFromDCS` Modifier to be able to read the magnet currents
from DCS. For this to work, we need to also read the maps themselves
from COOL and delay loading them to the first event. This should not
matter as this type of workflow would only be used for testing or debug
stream reprocessing, which both likely run with a single fork.

Also applied the same fixes to the "DB modifier" used during debug
stream reprocessing. But was not able to fully test this yet.

Closes ATR-24476.