TrigPSC: explicitly enable ROOT thread-safety
We relied on the HLTMPPU `MonSvcInfoService` library to enable ROOT's
general thread-safety. And due to a bug/feature we were loading this library
even when running in offline monitoring mode, so we never noticed this.
But if the library is not loaded, one immediately gets a crash when
accessing ROOT classes (typically the first one will be a histogram).

(in offline, this call is done in `RootStorageSvc`)