Ensure that closest parameters have a covariance matrix
MuonGetClosestParameters should check to make sure that the closest parameters have a covariance matrix if onlyUseMeasured is true.
However, at the moment, if the track doesn't reach to a chamber, so the closest parameters are the last ones, it doesn't check those.
Instead, due to a bug, it checks the first parameters.  This fix ensures that the correct parameters are checked so the covariance matrix is always present.