Fix for rare issue with the MuonTrackScoringTool
Tracks that are too poor quality to be scored are assigned a score of 0.
However, in rare cases, good tracks can also be given a score of 0, if the score assigned to the hits balances out the chi2.
This leads to the crash in ATLASRECTS-6670, in which the two tracks that are input to the MuonAmbiProcessor are almost identical.
Both are given a score of 0 and hence are rejected.
To prevent this, a score of 0 from the tool is changed to a score of 0.000001.