HLTSeeding: slightly reduce tau ROI size in phi
Hello,

We are currently using a tau ROI with an eta x phi half width of 0.4 x 0.4.
For phi, this is slightly larger than in Run2 where we were using Pi/8 ~ 0.393.
As discussed in ATR-23454, when using 0.4 instead of Pi/8, it seems that we are picking up an extra trigger tower in phi when unpacking LAr cells, resulting in LAr clusters outside the phi boundaries of the ROI.
This MR sets the phi half width to pi/8 to recover the Run2 behaviour.

Cheers,
Bertrand