GenericMonitoringTool: dynamically choose correct base class
Change the `GenericMonitoringTool` class to a function to return the
correct configurable base class. This is required to support mixed old
and new-style job options.

Also remove the "feature/hack" to always create a new configurable
instance. That was likely only needed because of a bug in the
`ConfigurableService` class fixed in !48419.

Relates to ATR-24495.