TauCombinedTES: add method to retrieve MVA tau energy resolution
Hello,

This MR implements a function in the TauCombinedTES tool to provide tau MVA energy resolution for the MET significance calculation.
So far, both in R21 and in R22, only the calo-only energy resolution was provided for analysis usage, but the calo-only Tau Energy Scale calibration is long superseded by the MVA TES calibration.
The old method has been removed to make sure it is no longer used.

The MVA energy resolution has been derived by Hanfei in ATLTAU-1740.
A new calibration file has been prepared (hadd of the new MVA parametrisations and the current calibration file CombinedTES_R22_Round2.5.root).
The new file is not deployed as new default in tauRec because this new functionality is not used in tau reconstruction, it's only an analysis feature.
Still, I have checked that the tau reconstruction is unchanged by this MR.
For completeness I have updated the file being used by CommonSmearingTool in TauAnalysisTools, but it won't change anything.

Cheers,
Bertrand