Fix for FastCaloSim running in ATLFASTIIF jobs
For whatever reason Fatras is creating particles that can cause very high `E_reco/E_true`
ratios in the `FastShowerCellBuilderTool::process_particle` method. When running FastCaloSim
with G4 in Run1 a maximum ratio of 3 or 4 worked without issues. We have already increased
this threshold once or twice in the last 2 years, but to allow further testing of ATLFASTIIF
the method has been changed to still return `StatusCode::SUCCESS` when the maximum ratio is
exceeded. The warning printouts are retained to allow follow ups on the FastChain side.

Should fix ATLASSIM-5434.