tauRecTools: fix spurious warnings
Hello,

As reported by Peter, the CombinedP4FromRecoTaus tool can issue spurious warnings as a consequence of !48449 .
The method returning the calo-only tau energy resolution (GetCaloResolution) is now superseded by the method returning the MVA tau energy resolution (GetMvaEnergyResolution), when used for the MET significance calculation.
So a warning was put in GetCaloResolution to make sure it would not be called.

I forgot that GetCaloResolution is internally used for other purposes in R21.2, and in that case the warnings are spurious.

So I am moving GetCaloResolution as a private method, and removed the warning.

Cheers,
Bertrand