CaloTowerxAODAlgoBase migrated to retrieving Calo DD manager from the Condition Store
In the old implementation, CaloTowerxAODAlgoBase used Calo DD manager at initialization
for building an index cache. The mechanism of building the cache did not seem to be
sensitive to Calo alignment changes. Hence, the simplest migration was to keep the cache
inside CaloTowerxAODAlgoBase, but turn it into a CxxUtils::CachedValue with lazy
initialization from execute().

The new implementation is expected to produce the same results as the old one (which had
not been listening to alignment callbacks anyway). However, if it gets determined that
the index cache building is indeed sensitive to alignment changes, then the cache must
be turned into a condition object (can be done in later MR-s, if needed).