Updating the derivation framework examples
This provides an update to the example DAOD formats in the derivation
framework, some of which were obsolete, and some of which were broken in
releae 22. The obsolete formats have been removed and the remaining ones
work in r22. This is in preparation for migrating the framework to the
ComponentAccumulator, which will begin with these example formats.