Remove dependence on RPC cabling from the RpcPadContainerCnv
The cabling is used only to define the size of the RpcPadContainer, which can just as easily be done using the RpcIdHelper.
The advantage of this approach is that the RpcIdHelper is available even if the muon reconstruction is not scheduled.
Presumably there aren't yet any cases in which the RpcPadContainerCnv needs to be run even though muon reco is not scheduled, but the track overlay will eventually need to be able to do this.