MuonCombinedBaseTools: Reproducibility fix.
The resolvedInDetCandidates vector is filled in part by iterating over a map
with pointer keys.  This means that the order of entries in this vector
can vary from run to run.  This was observed to lead to ordering differences
in the reconstruction output.  Add an additional sorting step to ensure
that results are identical from run to run.

The additional sort is keyed on track pt.  The ordering provided
by the existing SortInDetCandidates is too weak to help here.