limit warnings emitted from TauSelectionTool on older files
While working on some other tools I noticed that this warning emitted
by TauSelectionTool spams endlessly when running on an older file,
making it effectively unusable.  Now I don't know why the authors
decided not to let the exception escape here (which seems like the
right thing to do), but at the very least if we do this, the number of
printouts should be limited.

Also, by ATLAS convention, if the code continues on, this should be a
warning, not an error.