allows DIDType starting with "DIDType."
Not sure if this is a new development or just a corner case I hadn't
observed before, but a user reported an error that seems to be due to
an unanticipated DIDType string.  So I added a new set of names as an
alternative to the old set of names.