Updating PHYS(LITE/VAL) TCT
This MR updates the preExec of all PHYS(LITE/VAL) derivations TCT with the most recent one. It also updates the input file for the valid1PHYS with a more recent one produced with 22.0.47.1 and used also in PAV2. Note that the valid1PHYS test for 22.0-mc20/Athena/x86_64-centos7-gcc8-opt was crashing, as reported in https://its.cern.ch/jira/browse/ATLASG-1617, and it is fixed by the updated preExec.