remove uneccesary fullscan Roi flag
The fullscan flag is not needed here - it is really an abuse of the Roi descriptor to
use an Roi with different parameters where those parameters are used and still have it
as a fullscan Roi.

This removes the fullscan flag, and just uses a standard RoiDescriptor. This will add
around 1 ms to the SCT and Pixel RegionSelector calls, but should have otherwise no
impact.

If saving that 1 ms in the FS tracking instance taking ~ 600 ms is seen to be
important we can modify the SCT and Pixel data preparation to cache the full module list
so that the RS doesn';t need to be called at all, and then this will save around 7 ms
overall which would be a better solution.