Do not configure unused track summary tools.
In MT tracks have track summaries, so at many places like the detailed track
selector track summary tools are unused and not needed. Thus, they do not have
to be configured.

Also, increase severity to FATAL for cases in which a track passed to a
detailed track selector does not have a track summary and the selector
does not have a track summary tool.