TrigServices: Synchronize COOL folder updates across threads
Two independent changes to avoid (potential) race conditions when
updating COOL folders during the run:

1) Only decode the CTP fragment extra payload on LB changes. The
previous solution (checking on each event) was developed to be able to
update COOL folders for events in future LBs. But that is really not
needed. Checking the relevant folders to update on a LB change is
sufficient.

2) Ensure that no event is scheduled if a COOL update is ongoing.  This
should have minimal impact and will close a possible loop-hole when two
or more events at the start of a LB are scheduled, one of them
triggering a COOL folder reload and the other event possibly reading
outdated or corrupted data. Not entirely sure this is really needed, but
better to be on the safe side.

Relates to ATR-24864.