Fix all DataHandle(Key) holders that do not initialize the handles correctly (ATR-21690)
It is a widespread and reasonable, but incorrect, pattern to assume that if a handle(key) is not used then it should not be initialized.
In fact, this leaves the data dependency active in the scheduler, and no error will be thrown if the handle is never accessed.
Instead, the handle should be initialized(false) - this MR makes the change