Avoid deadlock between SGSvcImpl and IOVDbSvc.
Mitigated possible deadlock between SGSvcImpl and IOVDbSvc.
The global dblock applied in the IOVDbSvc led to a deadlock in
one case in which the order of the first call to SGSvcImpl and
IOVDbSvc was different, because the two services call each other.
This problem was solved by limiting the scope of the dblock.