CoreDumpSvc: add option to produce core file
Add `DumpCoreFile` property to try to produce a system core dump file as
a last action in the signal handling. Whether this works or not will
depend on the resource limits (e.g. `ulimit -c`).

Previously it was necessary to enable the "Seal" signal handler (e.g.
via `FatalHandler=639`) to get core files produced.