TileSimAlgs: Fix Tile digitization ART tests (ATLASSIM-5692, ATLASSIM-5695, ATLASSIM-5693)
The Tile sampling fractions in DB are not exactly (precisely)
the same as the ones which were used in JO before in old SF infrastructure.
This generates the difference in the output, but this difference is very small.

Tile digitization algorithms have been updated to use the same sampling fractions
as the ones which were used in JO to prove that the output is the same in regression tests.