TestTools: fail if reference file is empty after filtering
Change `nopost.sh` to fail if the original reference file is non-empty
but it becomes empty after applying the filtering rules (either default
or user provided patterns). This helps discovering cases where the user
(or changes in `post.sh`) accidentally filter the entire log and
reference file and cause tests to always succeed.