TrigValAlgs: enable thread-checker
Enable the thread-checker and apply some minimal fixes to
`TrigEDMChecker`. The code is still not thread-safe (and is not really
needed to be) but now at least the thread-hostile parts are clearly
marked as such and could be tackled in the future. Another option could
be to just delete the code that is able to dump the legacy EDM. But it
could also still be useful to inspect old AODs.