CITest: implement test dependencies on success
Add the `DEPENDS_SUCCESS` keyword to define test dependencies that
require successful completion of the dependee(s). Change all tests to
use this instead of `DEPENDS` and `REQUIRED_FILES`. The latter was not
reliable as even a failed test can create a (corrupt) AOD file for
example.

Also remove the dedicated parsing of the `DEPENDS` keyword as that is
now supported by `atlas_add_test` itself.