AthenaCommon: Be resilient when trying to preload Intel math library on ARM
The Intel math library (obviously) does not exist for ARM. But some test
jobs always preload it. In case we are running on ARM, just print a
warning that the library is not available. Note that on all other
platforms, not finding the library when it is requested does remain a
failure.

Closes ATLINFR-4503.