return a reference to an actual object in currentContext()
So far it had returned a reference to a `nullptr` to avoid anybody
somehow doing anything with it, but aparently by now gcc checks that
at run-time during debug builds, so this will now provide an actual
object and also provides a dummy class definition.  I added the later
as in the past I had seen people define their own version of
EventContext which would create problems when used with a concrete
object, and this at least gives the compiler a chance to catch that.