introduce std::string constructor for ReadHandle/WriteHandle
That allows to instantiate the handles directly from a string.  Right
now if a VarHandle gets instantiated with an std::string it will
create a temporary VarHandleKey which the VarHandle then references,
but which will be immediately destroyed, leading to a crash when the
handle is used.

This is only doing this for ReadHandle/WriteHandle, not for
ReadDecorHandle/WriteDecorHandle, as those don't have that capability
in the Athena implementation either and will crash there as well.