Remove Pixel and TRT PID computation from the TrackSummaryTool.
Also deprecate Pixel and TRT PID retrieval from the TrackSummary, and remove
support for PID selectors for Trk::Track from InDetTrackSelectionTool, since
this information is not computed and stored anymore for Trk::Track objects.
This information is not used in InDetTrackSelectionTool instances applied to
Trk::Tracks. So there are no consequences.

Support for the  TRT and Pixel PID information is dropped from the converter,
The definition of the persistent object used by the TrackSummary converter is
unchanged. In case the persistent object would contain TRT and Pixel PID
information, this information is now silently dropped. Since the TrackSummary
is only written out for the TrackParticleBase which had been replaced by the
xAOD::TrackParticle, this has no relevant consequences.