Add config flag for electron LH with HI and GSF NoPix, plus simplification
Add config flag for electron LH with HI and GSF NoPix, plus simplification

The config flag used for the HI menu was the same as the pp menu. This MR
gives each menu a separate config fag. It also allows for the option to
specify a different flag for NoPix with and without GSF. Finally, the logic
used to create the selector tools for the various combinations is
simplified.