Discard strips with small charge
sTGC strips with small charge are kept due to the near-neighbour logic.
With pileup, hits near the lower or upper time limit would created a
cluster of strips. However, not all these strips pass the selection
due to time smearing. Sometimes, only the strips with small charge
pass the selection, leading to a cluster without charge that breaks
the clusterizing method.