Component accumulator in DAODs - part 4: PHYSVAL
This MR adds a CA-compatible version of the PHYSVAL format. Aside of the new config file, it is also necessary to adjust the CMake file to account for the new python directory in the DerivationFrameworkPhysicsValidation package, and also to add the config fragment to the master list in DerivationFrameworkConfiguration.

Since this only impacts the CA version of the derivation framework, which isn't yet in use, this MR has no impact on any ongoing production or functionality.