first version of a SysReadSelectionHandle
The idea is to allow reading systematics varied selection decorations
with the existing ISelectionAccessor infrastructure and being fully
integrated with our systematics handles.

I updated all the algorithms to use this, but haven't yet updated the
configuration to use it, and also haven't added any specific tests for
this purpose.