Track overlay implementation part 3
Most of these changes are for the switch to using OverlayCopyAlgs to copy the containers created by the track overlay through the overlay digitization step.
This is implemented for both legacy and CA configuration.
This requires modifying the changes I previously made to the ID ROT converters to handle the fact that the PRD container names are hardcoded in the trans-to-pers conversion.
Luckily, this can be simplified, as a different name is now only needed at the end of the overlay step.
However, to make sure that the converter knows it is in this step, I have loaded the EventCnvTool into the overlay step (since the converters read this already, and I didn't want to make too many changes to them).
The remaining changes are to implement track overlay in the overlay step for the CA configuration.
If track overlay is on, then only the hard-scatter digitization is run, and not the rest of the ID overlay.