Fix TrigInDetAnalysis post processing
The chainnames were being mapped incorrectly. This fixes it so
tha the TrigInDetAnalysis code should work ok

A note to the reviewer:
All of the code in thios package is for the compilation of standalon
executables that are not tested with the atlas "required" CI
tests - only the build is relevant.