Component accumulator in DAODs - part 6: output stream settings
Draft: Component accumulator in DAODs - part 6: output stream settings

This MR applies the output stream settings (basket size, compression etc) in the CA version of the derivation framework that were applied in the old RecExCommon version. The settings are now applied in `PoolWriteConfig.py` which is the same place as AOD/ESD etc. This change has no impact on ongoing production as it only applies to the CA version of the DAODs, which aren't yet in use.