TrigT1NSWSimTools: fix thread-checker warnings in NTuple code
The NSW NTuple code can never be used in multi-threaded athena in its
current form. Instead of trying to fix it, make sure that this part of the
code is only used in single-threaded mode and mark the relevant
variables and functions as thread-safe.

But ideally this code should really be refactored or removed if not
needed anymore.

Relates to ATR-24727.